Revised scharfsandhi and ScharfSandhiArg
Added a second option to despace (Y) or not (N) for use with external sandhi (E).  Removed simple_sandhioptions.  For compound sandhi (C), the second option is irrelevant and may also be blank ('' or "" from the command line), but not absent.  E.g. python ScharfSandhiArg.py E N "tasmAt hasati"